requirements.txt
pytest==7.4.0

allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest-lazy-fixture==0.6.3

namedlist==1.8
collections-extended==2.0

setuptools==65.5.1